implemented value method with proper QVariant wrap

+++ src/demo.ts
@@ -5,7 +5,7 @@ const dateEdit = new QDateEdit();
 
 const settings = new QSettings("test1", "test2");
 console.log(settings.native);
-console.log(settings.value('heh'));
+console.log(settings.value('heh').toInt());
 settings.setValue('heh', 32);
 settings.sync();
 dateEdit.show();
